pyproject.toml
[project]
name = "jei_bot"
version = "0.1.0"
description = "Default template for PDM package"
readme = "README.md"
requires-python = "==3.12.*"
authors = [
    {name = "lzx20129002", email = "lzx20129002@163.com"},
]
dependencies = ["nonebot2[websockets]>=2.4.2", "nonebot-plugin-apscheduler>=0.5.0", "pillow>=11.1.0", "aiohttp>=3.11.16", "pyzbar>=0.1.9", "nonebot-adapter-onebot>=2.4.6", "aiomysql>=0.2.0", "cryptography>=44.0.2"]
license = {text = "MIT"}

[tool.nonebot]
adapters = [
    { name = "OneBot V11", module_name = "nonebot.adapters.onebot.v11" }
]
plugins = ["nonebot_plugin_apscheduler"]
plugin_dirs = ["src/plugins"]
builtin_plugins = []
[tool.pdm]
distribution = false
